Verifica se a tarefa existe antes de altgerar ou excluir uma tarefa

+++ src/routes.js
@@ -70,6 +70,12 @@ export const routes = [
           .end(JSON.stringify({ message: "description is required" }));
       }
 
+      const [task] = database.select("tasks", { id });
+
+      if (!task) {
+        return res.writeHead(404).end();
+      }
+
       database.update("tasks", id, {
         title,
         description,
@@ -84,6 +90,12 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params;
+
+      const [task] = database.select("tasks", { id });
+
+      if (!task) {
+        return res.writeHead(404).end();
+      }
 
       database.delete("tasks", id);
